requirements.txt
Babel==2.14.0
beautifulsoup4==4.7.1
dateparser==1.2.0
Jinja2==2.11
python-telegram-bot==12.8
PyYAML==6.0
requests==2.32
sh==1.12.14
SQLAlchemy==1.3.1
MarkupSafe==2.0.1
. # also run setup.py, with all dependencies pinned to a version
